backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
bcrypt
python-jose[cryptography]
httpx
pytest
